+++ src/classes/wp-post.class.ts
@@ -1,63 +1,82 @@
 import {WpUser} from "./wp-user.interface";
+import {IPost} from "./post.interface";
 
 export class WpPost {
 
-  constructor(public post) {
+  constructor(public post: IPost) {
   }
 
-  id = ():string=> {
+  get(property: string){
+    if(this.post[property]) return this.post[property];
+  }
+
+  id(): string {
     if (this.post.id) return this.post.id;
-  };
+  }
 
-  title = ():string=> {
+  slug(): string {
+    if (this.post.slug) return this.post.slug;
+  }
+
+  title(): string {
     if (this.post.title) return this.post.title.rendered;
-  };
+  }
 
-  content = ():string=> {
+  content(): string {
     if (this.post.content) return this.post.content.rendered;
-  };
+  }
 
-  excerpt = ():string => {
+  excerpt(): string {
     /** filter excerpt from the links */
     if (this.post.excerpt) return (<string>this.post.excerpt.rendered).replace(/<a\b[^>]*>(.*?)<\/a>/i, "");
-  };
+  }
 
-  date = ():string => {
+  date(): string {
     if (this.post.date) return this.post.date;
-  };
+  }
 
-  type = (): string =>{
+  link(): string {
+    if (this.post.link) return this.post.link;
+  }
+
+  type(): string {
     if (this.post.type) return this.post.type;
-  };
+  }
 
-  categories = ():any=> {
+  categories(): any {
     if (this.post._embedded && this.post._embedded['wp:term']) {
       return this.post._embedded['wp:term'][0];
     }
-  };
+  }
 
-  tags = ():any => {
+  format(): string {
+    if (this.post.format) {
+      return this.post.format;
+    }
+  }
+
+  tags(): any {
     if (this.post._embedded && this.post._embedded['wp:term'])
       return this.post._embedded['wp:term'][1];
-  };
+  }
 
-  author = ():WpUser => {
+  author(): WpUser {
     if (this.post._embedded)  return <WpUser>this.post._embedded.author;
-  };
+  }
 
   /** featuredMedia(): check if has featured image, return false | number */
-  featuredMedia = ():boolean | number => {
+  featuredMedia(): boolean | number {
     if (this.post.featured_media)  return +this.post.featured_media;
-  };
+  }
 
   /**
    * get post featured image url
    * @params {string} size -
    */
-  featuredImageUrl = (size:string):string => {
+  featuredImageUrl(size: string): string {
     if (this.featuredMedia() && this.post._embedded) {
 
-      var featuredImage = this.post._embedded['wp:featuredmedia'][0];
+      let featuredImage = this.post._embedded['wp:featuredmedia'][0];
       if (featuredImage) {
         if (featuredImage.media_details.sizes[size]) {
           return featuredImage.media_details.sizes[size].source_url;
@@ -72,32 +91,4 @@ export class WpPost {
 
 }
 
-interface thumbnail {
-  file:string;
-  height:string;
-  width:string;
-  mime_type:string;
-  source_url:string;
-}
 
-
-interface PostObject {
-  id:string;
-  slug:string;
-  name:string;
-  tags:any;
-  categories:any;
-  date:string;
-  modified:string;
-  type:string;
-  featured_media:number;
-  comment_status:boolean;
-  ping_status:boolean;
-  sticky:boolean;
-  format:string;
-  title:any;
-  content:any;
-  excerpt:any;
-  author:any;
-  _embedded:any;
-}