+++ src/service/collection/collection.interface.ts
@@ -1,13 +1,12 @@
 import {Observable} from 'rxjs/Observable';
-import {WpQueryArgs} from "../../helpers/wp-query.class";
 import {WpPagination} from "./collection.service";
 
 export interface CollectionInterface {
 
-  get(args?: WpQueryArgs): Observable<any>;
-  more(): Observable<any>;
-  next(): Observable<any>;
-  prev(): Observable<any>;
+  get(args?): Observable<any>;
+  more(): Observable<CollectionResponse>;
+  next(): Observable<CollectionResponse>;
+  prev(): Observable<CollectionResponse>;
 }
 
 export interface CollectionResponse {