+++ src/service/config/config.service.ts
@@ -1,52 +1,38 @@
-import {Injectable} from '@angular/core';
+import {Injectable, Optional} from '@angular/core';
 import {Headers} from "@angular/http";
-import {BehaviorSubject} from 'rxjs/BehaviorSubject';
 
 import {ConfigInterface} from './config.interface';
 
 @Injectable()
 export class ConfigService implements ConfigInterface {
 
-  /** base URL */
   baseUrl: string;
-  /** request timeout */
-  timeOut: number = 5000;
+  debug: boolean;
+  private authType: string;
+  private authKeys: string;
 
-  /** Subscribe to `loading` to get notified when the service is busy. */
-  loading: BehaviorSubject<boolean>;
-
-  /** Subscribe to `logs` to get notified whenever an error occurs. */
-  errors: BehaviorSubject<any>;
-
-  private _authType: AuthType;
-  private _authKeys: string;
-
-  constructor() {
-    this.loading = new BehaviorSubject<boolean>(false);
-    this.errors = new BehaviorSubject<any>(null);
+  constructor(@Optional() baseUrl: string) {
+    this.baseUrl = baseUrl;
   }
 
+  /** Get authentication key for HTTP Headers {cookies | basic auth} */
   getAuth(): Headers {
-    switch (this._authType) {
-      case AuthType.basic:
-        return basicHeaders(this._authKeys);
-      case AuthType.cookies:
-        return cookiesHeaders(this._authKeys);
-      default:
-        return undefined;
+    if (this.authKeys) {
+      if (this.authKeys.toLowerCase() === 'cookies') {
+        return cookiesHeaders(this.authKeys);
+      }
+      if (this.authKeys.toLowerCase() === 'basic') {
+        return basicHeaders(this.authKeys);
+      }
     }
+    return undefined;
   }
 
-  setAuth(keys: string, type: AuthType){
-    this._authKeys = keys;
-    this._authType = type;
+  setAuth(keys: string, type: string) {
+    this.authKeys = keys;
+    this.authType = type;
   }
 
-}
-
-export enum AuthType{
-  basic,
-  cookies
 }
 
 function basicHeaders(keys: string): Headers {