+++ src/service/model/model.interface.ts
@@ -1,12 +1,11 @@
 import {Observable} from 'rxjs/Observable';
-import {WpQueryArgs} from "../../helpers/wp-query.class";
 
 export interface ModelInterface {
 
-    get(id: number, args?: WpQueryArgs): Observable<any>;
-    add(body: any): Observable<any>;
-    update(id: number, body: any): Observable<any>;
-    delete(id: number): Observable<any>;
+    get(id: number, args?): Observable<ModelResponse>;
+    add(body: any): Observable<ModelResponse>;
+    update(id: number, body: any): Observable<ModelResponse>;
+    delete(id: number): Observable<ModelResponse>;
 }
 
 export interface ModelResponse {