+++ src/service/model/model.service.ts
@@ -1,9 +1,8 @@
 import {Injectable} from '@angular/core';
 import {Observable} from 'rxjs/Observable';
 
-import {WpHttp} from '../../helpers/wp-http.class';
-import {ModelInterface} from './model.interface';
-import {WpQueryArgs} from "../../helpers/wp-query.class";
+import {WpHttp} from '../../classes/wp-http.class';
+import {ModelInterface, ModelResponse} from './model.interface';
 /**
  * WpModel Service: Get/Add/Update/Delete single from WP API
  */
@@ -11,12 +10,12 @@ import {WpQueryArgs} from "../../helpers/wp-query.class";
 export class ModelService implements ModelInterface {
 
   /** Request Parameters  */
-  private WpQueryArgs: WpQueryArgs;
+  private args;
   private _id: number;
   private _body: any;
 
   constructor(private http: WpHttp, private endpoint: string) {
-    this.WpQueryArgs = new WpQueryArgs({});
+    this.args = {};
   }
 
   /**
@@ -25,10 +24,8 @@ export class ModelService implements ModelInterface {
    * @param args
    * @returns {Observable<Response>}
    */
-  get = (id?: number, args?: WpQueryArgs): Observable<any> => {
-    let reqId = (id) ? id : this._id;
-    let reqArgs = (args) ? args : this.WpQueryArgs;
-    return this.http.get(this.endpoint + reqId, reqArgs)
+  get(id: number, args?): Observable<ModelResponse> {
+    return this.http.get(this.endpoint + id, args)
       .map(res => {
         return { data: res.json() }
       })
@@ -41,7 +38,7 @@ export class ModelService implements ModelInterface {
    * @param body
    * @returns {Observable<Response>}
    */
-  add = (body?: any): Observable<any> => {
+  add(body?: any): Observable<ModelResponse>{
     let reqBody = (body) ? body : this._body;
     return this.http.post(this.endpoint, reqBody)
       .map(res => {
@@ -57,7 +54,7 @@ export class ModelService implements ModelInterface {
    * @param body
    * @returns {Observable<Response>}
    */
-  update = (id?: number, body?: any): Observable<any> => {
+  update(id?: number, body?): Observable<ModelResponse>{
     let reqId = (id) ? id : this._id;
     let reqBody = (body) ? body : this._body;
     return this.http.put(this.endpoint + reqId, reqBody)
@@ -73,9 +70,8 @@ export class ModelService implements ModelInterface {
    * @param id
    * @returns {Observable<Response>}
    */
-  delete = (id?: number): Observable<any> => {
-    let reqId = (id) ? id : this._id;
-    return this.http.delete(this.endpoint + reqId + "?force=true")
+  delete(id): Observable<ModelResponse> {
+    return this.http.delete(this.endpoint + id + "?force=true")
       .map(res => {
         return { data: res.json() }
       })
@@ -90,7 +86,7 @@ export class ModelService implements ModelInterface {
    * @returns {ModelService}
    */
 
-  body = (body: any): ModelService => {
+  body(body: any): ModelService {
     this._body = body;
     return this;
   };
@@ -100,7 +96,7 @@ export class ModelService implements ModelInterface {
    * @param id
    * @returns {ModelService}
    */
-  id = (id: number): ModelService => {
+  id(id: number): ModelService {
     this._id = id;
     return this;
   }